Atualizacao

+++ index.php
@@ -24,7 +24,15 @@ $app->get('/book/{id}', function (Request $request, Response $response) use ($ap
     $id = $route->getArgument('id');
     $entityManager = $this->get('em');
     $booksRepository = $entityManager->getRepository('App\Models\Entity\Book');
-    $book = $booksRepository->find($id);        
+    $book = $booksRepository->find($id); 
+    /**
+     * Verifica se existe um livro com a ID informada
+     */
+    if (!$book) {
+        $logger = $this->get('logger');
+        $logger->warning("Book {$id} Not Found");
+        throw new \Exception("Book not Found", 404);
+    }       
     $return = $response->withJson($book, 200)
         ->withHeader('Content-type', 'application/json');
     return $return;
@@ -50,6 +58,8 @@ $app->post('/book', function (Request $request, Response $response) use ($app) {
      */
     $entityManager->persist($book);
     $entityManager->flush();
+    $logger = $this->get('logger');
+    $logger->info('Book Created!', $book->getValues());
     $return = $response->withJson($book, 201)
         ->withHeader('Content-type', 'application/json');
     return $return;
@@ -71,6 +81,17 @@ $app->put('/book/{id}', function (Request $request, Response $response) use ($ap
     $booksRepository = $entityManager->getRepository('App\Models\Entity\Book');
     $book = $booksRepository->find($id);   
     /**
+     * Monolog Logger
+     */
+    $logger = $this->get('logger');
+    /**
+     * Verifica se existe um livro com a ID informada
+     */
+    if (!$book) {
+        $logger->warning("Book {$id} Not Found - Impossible to Update");        
+        throw new \Exception("Book not Found", 404);
+    }   
+    /**
      * Atualiza e Persiste o Livro com os parâmetros recebidos no request
      */
     $book->setName($request->getParam('name'))
@@ -80,7 +101,7 @@ $app->put('/book/{id}', function (Request $request, Response $response) use ($ap
      */
     $entityManager->persist($book);
     $entityManager->flush();        
-    
+    $logger->info("Book {$id} updated!", $book->getValues());
     $return = $response->withJson($book, 200)
         ->withHeader('Content-type', 'application/json');
     return $return;
@@ -96,18 +117,32 @@ $app->delete('/book/{id}', function (Request $request, Response $response) use (
     $route = $request->getAttribute('route');
     $id = $route->getArgument('id');
     /**
+     * Monolog Logger
+     */
+    $logger = $this->get('logger');
+    /**
      * Encontra o Livro no Banco
      */ 
     $entityManager = $this->get('em');
     $booksRepository = $entityManager->getRepository('App\Models\Entity\Book');
     $book = $booksRepository->find($id);   
     /**
+     * Verifica se existe um livro com a ID informada
+     */
+    if (!$book) {
+        $logger->info("Book {$id} not Found");
+        throw new \Exception("Book not Found", 404);
+    }     
+    /**
      * Remove a entidade
      */
     $entityManager->remove($book);
     $entityManager->flush(); 
+    $logger->info("Book {$id} deleted", $book->getValues());
     $return = $response->withJson(['msg' => "Deletando o livro {$id}"], 204)
         ->withHeader('Content-type', 'application/json');
     return $return;
 });
+
+
 $app->run();
